requirements.txt
astroid==3.3.9
contourpy==1.3.2
cycler==0.12.1
dill==0.3.9
fonttools==4.57.0
isort==6.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pylint==3.3.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tomlkit==0.13.2
tzdata==2025.2
